REVERTME: fpga: dfl: init needs_suppliers after disabling sriov
device_links_purge() is called by device_del(). It removes
the device from several lists including consumer, suppliers,
and needs_suppliers. In the case of needs_suppliers, it uses
list_del(), which places poison values in the next and
previous pointers.

When using SRIOV, the fpga dfl driver uses
platform_device_del() to detach a device, and
platform_device_add() to reattach the device.  In this
context device_add() fails because of the poisoned
needs_suppliers pointers.  The other lists (consumers and
suppliers) are in the initialized state and are fine.

Comments in the mainline code for device_add() state: "Do not call
this routine or device_register() more than once for any device
structure." The current DFL driver violates this and needs to
be reworked to create a new device structure when reattaching
after SRIOV.

This is a temporary, short term fix which reinitializes the
needs_suppliers list prior to calling device_add() when
reattaching the device.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>