fpga: sec-mgr: do not require line-return for image filename
The filename sysfs entry that is used to start a secure update
was requiring a line-return at the end of the filename (e.g.
echo IMAGE_FILE > filename). This is incorrect, so the line-return
is being removed. To start an update directly with sysfs, do
"echo -n IMAGE_FILE > filename".

Signed-off-by: Russ Weight <russell.h.weight@intel.com>