memory: dfl-emif: add hard-coded dynamic module alias
Module aliases are generated by scripts/mod/file2alias based on the id
passed to MODULE_DEVICE_TABLE, which in our case is 'dfl'.

When building backported dfl modules the file2alias from the host kernel
is used, in which case the 'dfl' id is unknown. This leads to modules
not being loaded when devices are discovered, and hence lacking
functionality.

Fix this by adding a hard-coded module alias, as we cannot patch the
host file2alias.

Signed-off-by: Martin Hundebøll <mhu@silicom.dk>