rpm: remove use of the environment variable BACKPORT_VERSION
BACKPORT_VERSION can be used give arbitrary versions to the rpm.
The control of the version should be contained in the spec file
and be unique.

The version is
<os_branch> + '_' + <lts_tag>

For this branch, os_branch is rhel8 and the current lts_tag is v5.10.75
so the version is rhel8_v5.10.75

Since this is the second release under v5.10.75, the release is
is incremented to 2.  The the lts_tag changes, the release will reset
to 1.

When rhel9/ is started os_branch will change to rhel9

Signed-off-by: Tom Rix <trix@redhat.com>