build: Move functions from rpm spec to dkms
Move some pre/post operations (e.g. modprobe and rmmod) from the RPM
spec file into the dkms pre/post scripts. This is probably more correct,
as it better supports stand-alone dkms commands. It also allows these
functions to be shared with the debian package implementation.

Also moved dkms.conf to dkms.conf.in to make it more clear that the
file is being preprocessed to generate the final dkms.conf. Change
the default PACKAGE_VERSION in dkms.conf.in from "5.11" to "PKGVER"
to make it more obvious that we are replacing the string during
preprocessing.

Add the LICENSE file to the generated RPM package payload.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>