build: Fix dkms package names in RPM spec
"-dkms" was added the package name in the RPM spec file, but that doesn't
mean that has to be used for the dkms package name or for the temorary
directories. Remove "-dkms" from the dkms package names in order to
restore consistent naming. This fixes a problem the prevents the RPM
from installing correctly.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>