fpga: dfl: Add GUID support in mod device table
In the modalias table, dfl devices use the type id and feature id to
create the dfl device string entry "dfl:t0000f0009*". This patch adds
support for the use of a GUID for dfl devices. Device drivers may match
on type id and feature id, on the GUID, or on a combination of the type id,
feature id and GUID.

If the feature id and GUID are both non-zero, then the modalias string is:
"dfl:t0000f0009g{03020100-0504-0706-0809-0A0B0C0D0E0F}*"
If the feature id is zero and the GUID is non-zero, then the string is:
"dfl:t*f*g{03020100-0504-0706-0809-0A0B0C0D0E0F}*"
If the feature id is non-zero and the GUID is zero, then the string is
"dfl:t0000f0009*"

Followed the concept of the wmi device ID to define the GUID in string
format and reuse the existing GUID infrastructure.

Signed-off-by: Basheer Ahmed Muddebihal <basheer.ahmed.muddebihal@linux.intel.com>