mfd: intel-m10-bmc: no handshake_regs for N6000 and C6100
The N6000 and C6100 do not need to intercept handshake register accesses
during a secure update in the way that the N3000 and D5005 devices do.
Therefore they do not need to define regmap_range arrays. This change
removes the n6000_fw_handshake_regs[] and m10bmc_c6100_bmc_subdevs[]
arrays and creates an empty null_fw_handshake_regs[] array to be used
for these and similar devices. Some of the comment text is also reduced
and simplified.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>