backport: Adapt image-load for latest secure update driver
The secure update driver has been accepted upstream and uses the new
firmware-upload functionality of the firmware loader. Unfortunately, the
firmware-upload functionality lives in the kernel proper, which creates
a problem for out-of-tree drivers: if the base kernel does not include
the firmware-upload functionality, then the out-of-tree driver must carry
it's own solution. To handle this case, the image-load driver is being
modified with two goals in mind:

(1) Minimize changes to the secure-update driver
(2) Require no changes to user-space code

In order to accomplish these goals, a significant rewrite of the FPGA
image load driver is required. As much as possible, code from the
firmware-loader implementation is used to simplify future maintenance
of the driver.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>