mfd: intel-m10-bmc: Change staging size to a variable
The size of the staging area in FLASH for FPGA updates is dependent on the
size of the FPGA. Currently, the staging size is defined as a constant.
Larger FPGAs are coming soon and it will soon be necessary to support
different sizes for the staging area. Add a new staging_size member to the
m10bmc_csr structure to support a variable staging size.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>
(cherry picked from commit e9eb96934ecde902bf490c61dab3728edd83bd18)