fpga: dfl: Prefer GUID match when binding dfl drivers
With the introduction of GUIDs as DFL device identifiers, a DFL driver may
be selected based on a GUID match or based on a type/feature ID match. A
driver that matches based on GUID is written specifically for that device.
A driver that matches based on type/feature ID is likely to be a more
generalized driver that will support the device but may not enable all
functionality. This change will prevent a DFL device from binding to a
driver based on type/feature ID if a driver for the GUID is present in
the system.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>
Signed-off-by: Basheer Ahmed Muddebihal <basheer.ahmed.muddebihal@linux.intel.com>