fpga: dfl: Add feature device driver match by GUID
This patch introduces changes to enhance the identification and
matching of features in the Device Feature List (DFL) using GUID
(Globally Unique Identifier). The use of GUID allows for an infinite
pool of feature IDs, and decentralized allocation, and enables the
matching of DFL devices with their respective drivers based on GUID.
This mechanism allows customers to match their custom IP with the
desired or required driver.

Changes made in this patch:

1. Modified the UEVENT string to reflect GUID matching, ensuring
compatibility with the updated identification method.
2. Added system entries to display the GUID in Little Endian format,
facilitating the interpretation of the GUID value.
3. Included wrapper functions for GUID matching and Feature ID matching,
enabling the comparison and matching of GUIDs and feature.

Signed-off-by: Basheer Ahmed Muddebihal <basheer.ahmed.muddebihal@linux.intel.com>