backport: compatibility: Allow parameter size zero
The DFHv1 parameter block architecture was modified during upstreaming.
Previously a parameter size of zero marked the end of the list. In the
accepted implementation, the size must not be zero and the end of list
is marked with a flag.

This change is made to support both implementations by checking for a
parameter size of zero, accepting it as the end of list, and substituting
the zero size for a size of 16 bytes.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>