build: Handle install conflict with pre-existing drivers
When installing on a new system, the dkms package:
(1) Builds the binaries
(2) Stores the binary modules into the appropriate directory under
    /lib/modules/`uname -r`
(3) Unloads (rmmod) pre-existing driver modules
(4) Loads the newly built driver modules (modprobe dfl_pci)

Step (3) is problematic specifically in the case there the running kernel
includes another version of the same drivers and those driver modules are
already loaded in the running kernel. The module names and the list of
modules known to the driver package may not match exactly what is loaded
in the kernel. After the driver package attempts to rmmod the currently
loaded modules, there could be related modules that remain loaded. The
driver package then loads new modules alongside the old. This results in
kernel symbol errors in the kernel log file. In some cases the user
experiences undefined/confusing behavior.

The simplest and most reliable solution to this problem is to reboot
the system if there is potential for a conflict. The boot process
will guarantee a consistent state using only the new driver modules.

To prevent a conflict in driver versions, the following changes are made:
(1) Remove the preinstall script (which does rmmod). This means that no
    attempt will be made to unload an older version of the drivers from
    the running kernel (assuming it exists).
(2) Change the postinstall script to instruct the user to reboot the system
    _if_ there is another version the dlf_pci module already loaded in the
    running kernel. If there is no pre-existing instance of dfl_pci loaded,
    then they newly installed drivers will be loaded into the running kernel
    (as they have been historically).

Signed-off-by: Russ Weight <russell.h.weight@intel.com>