backport: Avoid function name conflict in rhel8.8
Fix a name conflict for RHEL 8.8 that was introduced by:

64772ab52f7e backport: Avoid memcpy_fromio() issues on older kernels

The above patch added a new function, unrolled_memcpy_fromio(), to be used
to avoid performance issues with older kernels. Unfortunately, this
same function name is also present in RHEL 8.8 with a different definition.
Rename the function to backport_memcpy_fromio() to avoid the conflict.

The backport_memcpy_fromio() function definition is also simplified.

Signed-off-by: Russ Weight <russell.h.weight@intel.com>