net: phy: qsfp-mem: Handle NACK from QSFP module
Depending on the manufacturer of particular QSFP cables,
the time a QSFP cable responds to a command after a
reset varies. If the cable is not ready, it will respond
with a NACK that must be cleared.

Co-developed-by: Yoshiaki Saito <yoshiaki.saito@intel.com>
Signed-off-by: Matthew Gerlach <matthew.gerlach@linux.intel.com>
(cherry picked from commit 3a0e610ce9e4b6b0a5b90207b136429de6ee63b0)