Makefile: support builtin fpga-{mgr,bridge,region} modules
The Intel SoCFPGA GSRD (Golden Software Reference Design) Yocto image
is by default built with CONFIG_FPGA=y (but CONFIG_FPGA_BRIDGE=m and
CONFIG_FPGA_REGION=m), which causes an unresolvable symbol conflict
between the kernel image and the fpga-mgr backport module.

Assuming that the kernel version is recent enough not to require any
updates to fpga-{mgr,bridge,region}, do not build these modules when
CONFIG_FPGA=y. Do not consider CONFIG_FPGA_BRIDGE and CONFIG_FPGA_REGION
to avoid a version mix between kernel-provided and backport modules.

This is not a complete solution and may result in unknown failures when
the builtin kernel modules are too old, but it does happen to work with
Intel SoCFPGA GSRD which is currently based on Linux 6.1.55 (LTS).

Link: https://github.com/altera-opensource/gsrd-socfpga
Link: https://github.com/altera-opensource/linux-socfpga
Signed-off-by: Peter Colberg <peter.colberg@intel.com>