.github: workflows: schedule weekly container and module builds
Build modules with latest kernels, e.g., as part of security updates.

The minutes are more or less chosen at random, with sufficient wait
time between building the containers followed by building the modules
using the new containers. The hour falls into the middle of a workday
such that a potential breakage does not need to be addressed as the
first task of the day, but may still be resolved until end of the day.

Signed-off-by: Peter Colberg <peter.colberg@intel.com>